requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioice==0.10.1
aiortc==1.12.0
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.1
attrs==24.1.0
autobahn==19.11.2
Automat==25.4.16
av==14.4.0
azure-core==1.34.0
azure-identity==1.23.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.2.1
cloudevents==1.11.1
colorama==0.4.6
comm==0.2.3
constantly==23.10.4
contourpy==1.3.3
cryptography==45.0.2
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
distlib==0.3.8
distro==1.9.0
Django==5.0.7
django-debug-toolbar==4.4.6
django-otp==1.6.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
executing==2.2.1
faiss-cpu==1.13.2
fastapi==0.115.13
filelock==3.15.4
Flask==3.1.1
Flask-WTF==1.2.2
fonttools==4.62.1
frozenlist==1.6.0
google-crc32c==1.7.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperlink==21.0.0
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.6.1
incremental==24.7.2
iniconfig==2.1.0
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiteconnect==5.0.1
kiwisolver==1.5.0
langchain==1.2.12
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.20
langchain-openai==1.1.11
langchain-text-splitters==1.1.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.20
lazy-object-proxy==1.11.0
limits==5.8.0
MarkupSafe==3.0.2
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
more-itertools==10.7.0
msal==1.32.3
msal-extensions==1.3.1
multidict==6.4.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.2.6
openai==2.29.0
openapi-core==0.19.5
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
orjson==3.11.7
ormsgpack==1.12.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.0
pandas_ta==0.3.14b0
parse==1.20.2
parso==0.8.5
pathable==0.4.4
pillow==12.1.1
pipenv==2024.0.1
platformdirs==4.2.2
pluggy==1.6.0
prance==25.4.8.0
prompt_toolkit==3.0.52
propcache==0.3.1
protobuf==6.31.0
psutil==6.1.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybars4==0.9.13
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.13.1
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.18.0
PyJWT==2.10.1
pylibsrtp==0.12.0
PyMeta3==0.5.1
pyOpenSSL==25.1.0
pyparsing==3.3.2
Pypubsub==4.0.3
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pywin32==308
PyYAML==6.0.2
pyzmq==27.1.0
redis==7.3.0
referencing==0.36.2
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
robotframework==7.1.1
robotframework-ride==2.1.1
rpds-py==0.25.1
rsa==4.9.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scipy==1.15.3
selenium==4.23.1
semantic-kernel==1.31.0
service-identity==24.2.0
setuptools==71.1.0
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.46.2
ta==0.11.0
tenacity==9.1.4
tiktoken==0.12.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
trio==0.26.1
trio-websocket==0.11.1
Twisted==25.5.0
txaio==25.6.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2024.1
urllib3==2.2.2
uuid_utils==0.14.1
uvicorn==0.34.3
virtualenv==20.26.3
wcwidth==0.5.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.1
wheel==0.44.0
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
wxPython==4.2.2
xxhash==3.6.0
yarl==1.20.0
zipp==3.21.0
zope.interface==7.2
zstandard==0.25.0
